Each user can sign in with Google and it will get the list of the user's calendars.

+++ database/migrations/1481148628058_user_accounts.js
@@ -14,7 +14,7 @@ class UsersTableSchema extends Schema {
       table.string('name').defaultTo('Anonymous')
       table.enu('type', ['worker', 'user'])
       table.text('token').nullable()
-      table.integer('expiry_date').nullable()
+      table.bigInteger('expiry_date').nullable()
       table.timestamps()
     })
   }
